+++ app/suite/clients/page.jsx
@@ -1,0 +1,12 @@
+import { Box, Toolbar, Typography } from "@mui/material";
+import React from "react";
+
+function Clients() {
+  return (
+    <>
+      <Typography>Clients</Typography>
+    </>
+  );
+}
+
+export default Clients;
